Updating clients saving methods
Now don't remove automatically all victims folder in action=cleanit
We check for each victim folder (IP) if the folder got updated by action=writeme
If not folder age in timestamp will be older.If the timestamp of the folder is 5s older than the Actual date. 
We delete the folder
This helps optimisation, as sometimes stub won't receive instruction because they will try to read it while the folder isn't written after delete.